feat: allow debugging a single nested test at any level
